requirements.txt
asgiref==3.4.1
Django==3.2
django-cors-headers==3.10.1
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
